requirements-web.txt
duckduckgo-search
requests
beautifulsoup4
readability-lxml
lxml-html-clean
html5lib